Refactoring of the disassembler completed.
The disassembler has been refactored to use the new identify_ins()
function. The main.c file has been renamed to disassembler.c which
simple calls chip8disasm().

opcodes.h has been added, and contains an enum for the 35 instructions
for Chip8.

CMakeLists.txt files have also been added, and are used by Cmake to
generate the Makefiles.

+++ src/utility/utility.h
@@ -1,7 +1,8 @@
 #ifndef __GUARD_UTILITY__
 #define __GUARD_UTILITY__
+#include <stdlib.h>
 
-size_t read_rom(const char* rom_file, char *data, size_t data_size);
+int read_rom(const char* rom_file, char *data, size_t data_size);
 char nibble(short word, unsigned int which);
 char low_byte(short word);
 short get_addr(short opcode);
